+++ changelog.md
@@ -14,6 +14,9 @@ For now, please use all minor versions as major versions and each build version 
 ### Changed
 - `Organisation`'s profile changed from an array of profiles to a single `Profile`
 
+### Removed
+- `organisation` from `ProductTag`
+
 ## [0.4.0]
 ### Removed
 - `amount` from `CreateCreditCardPaymentRequest` & `CreateCashOnDeliveryPaymentRequest`